Refactor to use db object
