+++ plugins/@grouparoo/demo/src/bin/grouparoo/demo/dataPurchases.ts
@@ -1,15 +1,16 @@
 import { CLI, log } from "actionhero";
-import { users, purchases } from "../../../postgres/sample_data";
+import { users, purchases } from "../../../util/sample_data";
 import { loadConfigFiles } from "../../../util/configFiles";
 import { events } from "../../../events/events";
 import { init, finalize } from "../../../util/shared";
+import Connection from "../../../postgres/connection";
 
 export class Console extends CLI {
   constructor() {
     super();
     this.name = "demo-data-purchases";
     this.description =
-      "Load eCommerce users and purchases into a source database.";
+      "Load eCommerce users and purchases into a source database and create properties including events.";
     this.inputs = {
       scale: { required: false, default: "1" },
     };
@@ -17,14 +18,15 @@ export class Console extends CLI {
 
   async run({ params }) {
     const scale = parseInt(params.scale);
+    const db = new Connection();
     if (scale > 1) log(`Using scale = ${params.scale}`);
 
     await init({ reset: true });
-    await users({ scale });
-    await purchases({ scale });
+    await users(db, { scale });
+    await purchases(db, { scale });
     await events({ scale });
-    await loadConfigFiles("setup");
-    await loadConfigFiles("purchases");
+    await loadConfigFiles(null, "setup");
+    await loadConfigFiles(db, "purchases");
     await finalize();
     return true;
   }
