+++ plugins/@grouparoo/demo/src/bin/grouparoo/demo/setup.ts
@@ -11,7 +11,7 @@ export class Console extends CLI {
 
   async run() {
     await init({ reset: true });
-    await loadConfigFiles("setup");
+    await loadConfigFiles(null, "setup");
     return true;
   }
 }
