+++ plugins/@grouparoo/demo/src/postgres/connection.ts
@@ -1,19 +1,34 @@
 import { Client } from "pg";
-import { log, execSync, userCreatedAt, numberOfUsers } from "../util/shared";
+import { log } from "../util/shared";
 import { api, config } from "actionhero";
-import parse from "csv-parse/lib/sync";
-import fs from "fs";
-import path from "path";
+import Connection from "../util/connection";
 
 export const SCHEMA_NAME = "demo";
 
-export function readCsvTable(tableName) {
-  const filePath = path.resolve(
-    path.join(__dirname, "..", "..", "data", `${tableName}.csv`)
-  );
-  const rows = parse(fs.readFileSync(filePath), { columns: true });
-  return rows;
-}
+const TYPES = {
+  users: {
+    id: "INT NOT NULL PRIMARY KEY",
+    first_name: "VARCHAR(191) NOT NULL",
+    last_name: "VARCHAR(191) NOT NULL",
+    email: "VARCHAR(191) NOT NULL",
+    gender: "VARCHAR(191)",
+    ip_address: "VARCHAR(191)",
+    avatar: "VARCHAR(191)",
+    language: "VARCHAR(191)",
+    deactivated: "BOOLEAN",
+    created_at: "TIMESTAMP NOT NULL",
+    updated_at: "TIMESTAMP NOT NULL",
+  },
+  purchases: {
+    id: "INT NOT NULL PRIMARY KEY",
+    user_id: "INT NOT NULL",
+    item: "VARCHAR(191) NOT NULL",
+    category: "VARCHAR(191) NOT NULL",
+    price: "DECIMAL",
+    state: "VARCHAR(191)",
+    created_at: "TIMESTAMP NOT NULL",
+  },
+};
 
 function findConfig() {
   const connectionURL = process.env.DEMO_DATABASE_URL;
@@ -60,20 +75,22 @@ function findConfig() {
   return clientConfig;
 }
 
-export default class Postgres {
+export default class Postgres extends Connection {
   client: Client;
   config: { [key: string]: any };
   constructor(schema = SCHEMA_NAME) {
+    super();
     this.config = Object.assign({}, findConfig(), { schema });
     this.client = null;
   }
 
-  getConfig(): { [key: string]: any } {
-    return this.config;
-  }
-
-  async execSync(level, command) {
-    execSync(level, command);
+  getAppOptions() {
+    const { host, port, database, schema, user, password } = this.config;
+    const appOptions = { host, port, database, schema, user, password };
+    for (const key in appOptions) {
+      appOptions[key] = (appOptions[key] || "").toString();
+    }
+    return appOptions;
   }
 
   async disconnect() {
@@ -95,6 +112,48 @@ export default class Postgres {
     return this.client;
   }
 
+  async createTable(tableName: string, userId: string, keys: string[]) {
+    const sqlTable = `${this.config.schema}."${tableName}"`;
+
+    const typeData = TYPES[tableName];
+    if (!typeData) {
+      throw new Error(`Unknown tableName: ${tableName}`);
+    }
+    const columnTypes = [];
+    for (const key of keys) {
+      const type = typeData[key];
+      if (!type) {
+        throw new Error(`Unknown column for ${tableName}: ${key}`);
+      }
+      columnTypes.push(`${key} ${type}`);
+    }
+
+    await this.query(1, `DROP TABLE IF EXISTS ${sqlTable}`);
+
+    const createQuery = `CREATE TABLE ${sqlTable} (${columnTypes.join(", ")})`;
+    await this.query(1, createQuery);
+
+    if (userId !== "id") {
+      const indexQuery = `CREATE INDEX "${tableName}_userId" ON ${sqlTable} ("${userId}");`;
+      await this.query(1, indexQuery);
+    }
+  }
+
+  async insertRow(
+    tableName: string,
+    keys: string[],
+    row: { [key: string]: string | number | Date }
+  ) {
+    const sqlTable = `${this.config.schema}."${tableName}"`;
+    const columnNames = keys.join(", ");
+    const variables = keys.map((key, index) => "$" + (index + 1));
+    const insertQuery = `INSERT INTO ${sqlTable} (${columnNames}) VALUES (${variables})`;
+
+    const values: Array<any> = keys.map((key) => getValue(row, key));
+    console.log(values.join(", "));
+    await this.query(2, insertQuery, values);
+  }
+
   async query(level, sql, params = null) {
     log(level, sql, params);
     const client = await this.connect();
@@ -103,111 +162,9 @@ export default class Postgres {
     }
     return client.query(sql);
   }
-
-  async createCsvTable(
-    tableName: string,
-    userId: string,
-    types: any,
-    createdAt: boolean,
-    updatedAt: boolean,
-    scale: number = 0
-  ) {
-    const sqlTable = `${this.config.schema}."${tableName}"`;
-    await this.query(1, `DROP TABLE IF EXISTS ${sqlTable}`);
-
-    if (scale < 1) {
-      scale = 1;
-    }
-    log(1, `Adding ${tableName}`);
-    // read from data file
-    const rows = readCsvTable(tableName);
-    const csvKeys = Object.keys(rows[0]);
-    const typeKeys = Object.keys(types);
-    const diff = differenceOf2Arrays(csvKeys, typeKeys);
-    if (diff.length > 0) {
-      throw `${tableName} csv and types do not match: ${diff.toString()}`;
-    }
-
-    if (createdAt) {
-      typeKeys.push("created_at");
-      types.created_at = "TIMESTAMP NOT NULL";
-    }
-    if (updatedAt) {
-      typeKeys.push("updated_at");
-      types.updated_at = "TIMESTAMP NOT NULL";
-    }
-
-    const columnTypes = typeKeys.map((key) => `${key} ${types[key]}`);
-    const createQuery = `CREATE TABLE ${sqlTable} (${columnTypes.join(", ")})`;
-    await this.query(1, createQuery);
-
-    if (userId !== "id") {
-      const indexQuery = `CREATE INDEX "${tableName}_userId" ON ${sqlTable} ("${userId}");`;
-      await this.query(1, indexQuery);
-    }
-
-    const db = this;
-    const fillTable = async function (page: number = 0) {
-      log(2, `   Page ${page}`);
-      const perPage = rows.length;
-      const columnNames = typeKeys.join(", ");
-      const variables = typeKeys.map((key, index) => "$" + (index + 1));
-      const insertQuery = `INSERT INTO ${sqlTable} (${columnNames}) VALUES (${variables})`;
-      for (const fileRow of rows) {
-        const row = Object.assign({}, fileRow);
-        row.id = parseInt(row.id);
-        if (row.id <= 0) {
-          throw new Error(`no id column on ${tableName}`);
-        }
-        row[userId] = parseInt(row[userId]);
-        if (row[userId] <= 0) {
-          throw new Error(`no ${userId} column on ${tableName}`);
-        }
-
-        const now = new Date().getTime();
-        let generatedCreateAt = userCreatedAt(row[userId]);
-        let creationAgo = now - generatedCreateAt.getTime();
-        if (tableName !== "users") {
-          creationAgo = Math.random() * creationAgo;
-          const updatedMilli = now - creationAgo;
-          generatedCreateAt = new Date(updatedMilli);
-        }
-
-        if (createdAt) {
-          row.created_at = new Date(generatedCreateAt);
-        }
-        if (updatedAt) {
-          // sometime after that
-          const updatedAgo = creationAgo * Math.random();
-          const updatedMilli = now - updatedAgo;
-          row.updated_at = new Date(updatedMilli);
-        }
-
-        if (page > 1) {
-          row.id = perPage * page + row.id;
-          if (tableName !== "users") {
-            row[userId] = numberOfUsers * page + row[userId];
-          }
-          if (row.email) {
-            row.email = `${page}${row.email}`;
-          }
-        }
-
-        const values: Array<any> = typeKeys.map((key) =>
-          getValue(row, key, types)
-        );
-        console.log(values.join(", "));
-        await db.query(2, insertQuery, values);
-      }
-    };
-
-    for (let page = 1; page <= scale; page++) {
-      await fillTable(page);
-    }
-  }
 }
 
-function getValue(row, key, types) {
+function getValue(row, key) {
   const value = row[key];
   if (!value) {
     return null;
@@ -216,16 +173,4 @@ function getValue(row, key, types) {
     return null;
   }
   return value;
-}
-
-function differenceOf2Arrays(array1, array2) {
-  const temp = [];
-
-  for (var i in array1) {
-    if (!array2.includes(array1[i])) temp.push(array1[i]);
-  }
-  for (i in array2) {
-    if (!array1.includes(array2[i])) temp.push(array2[i]);
-  }
-  return temp.sort((a, b) => a - b);
 }
